pyproject.toml
[project]
name = "wellness-agent-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "gradio>=6.9.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langgraph>=1.1.2",
    "matplotlib>=3.10.8",
    "nbformat>=4.2.0",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
]
